agent-service/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-service"
version = "0.1.0"
description = "Generative agent service for PJ conversational assistant"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.116.1",
  "uvicorn[standard]>=0.35.0",
  "pydantic>=2.11.7",
  "langgraph>=0.6.6",
  "langchain-core>=0.3.74",
  "prometheus-client>=0.22.1",
  "numpy>=2.0.2",
  "scikit-learn>=1.6.1",
  "joblib>=1.5.2",
  "opentelemetry-api>=1.38.0",
  "opentelemetry-sdk>=1.38.0",
  "opentelemetry-exporter-otlp>=1.38.0",
  "opentelemetry-instrumentation-fastapi>=0.59b0",
  "openai>=1.108.1",
  "langfuse>=2.60.0,<3.0.0",
  "tiktoken>=0.11.0",
  "eval-type-backport>=0.2.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.2",
  "pytest-asyncio>=1.2.0",
  "httpx>=0.28.1",
  "jsonschema>=4.25.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
